+++ Patches/JewelSpawnSystemPatch.cs
@@ -19,7 +19,7 @@ internal static class JewelSpawnSystemPatch
     static readonly System.Random _random = new();
     // static Unity.Mathematics.Random _unityRandom = new();
 
-    const double SKEW_FACTOR = 0.25;
+    const double SKEW_FACTOR = 0.1;
 
     static readonly bool _extraRecipes = ConfigService.ExtraRecipes;
 
@@ -84,7 +84,7 @@ internal static class JewelSpawnSystemPatch
                     for (int i = 4; i < 8 && assignedMods < newSpellMods.Count; i++)
                     {
                         PrefabGUID spellModPrefabGUID = newSpellMods[assignedMods];
-                        float powerValue = GetPowerValueForSpellMod(spellModPrefabGUID);
+                        float powerValue = GetRandomPower();
                         // float powerValue = Mathf.Clamp((float)_random.NextDouble(), 0.5f, 1f);
 
                         switch (i)
@@ -143,7 +143,6 @@ internal static class JewelSpawnSystemPatch
                     // SpellModSyncSystemServer.OnUpdate();
                     // __instance.InitializeJewelOnSpawn(entity, ref _unityRandom);
                     // __instance.InitializeSpawnedJewel(uninitializedJewel, false);
-
                     // Core.Log.LogWarning($"JewelSpawnSystemPatch - Added {newSpellMods.Count} spell mods to {prefabGuid.GetPrefabName()}");
                 }
             }
@@ -338,7 +337,7 @@ internal static class JewelSpawnSystemPatch
             (list[n], list[k]) = (list[k], list[n]);
         }
     }
-    static float GetPowerValueForSpellMod(PrefabGUID spellModPrefabGUID)
+    static float GetRandomPower()
     {
         // double minPower = powerRange.Value.Min;
         // double maxPower = powerRange.Value.Max;
@@ -373,7 +372,7 @@ internal static class JewelSpawnSystemPatch
         { new(189403977), new() { new(459492812), new(786676751), new(515468772), new(-1106879810), new(-2026740129), new(2115999081), new(-1721922606), new(-2009288107) } }, // sanguine coil
         { new(-880131926), new() { new(-1144993512), new(411514116), new(-218122346), new(-1967214301), new(-111114882), new(1439297485), new(-1967899075), new(-2009288107) } }, // shadowbolt
         { new(305230608), new() { new(626026650), new(1855739816), new(156877668), new(1384658374), new(255266111), new(-1430581265) } }, // veil of blood
-        { new(-1380116221), new() { new(-728088234), new(189108909), new(-691013912), new(159297406), new(-775157186), new(389507623), new(-593275986), new(-1925490558), new(890139957) } }, // carrion swarm
+        { new(-1380116221), new() { new(-1547199412), new(-884299127), new(-2020998085), new(-1837470543), new(1940553981), new(1915162299) } }, // carrion swarm
         // chaos
         { new(1575317901), new() { new(-648008702), new(-68573491), new(-960235388), new(2113057383), new(1439297485), new(-1772665607) } }, // aftershock
         { new(-1016145613), new() { new(-547116142), new(-1611128617), new(1906516980), new(1600880528), new(-1251505269), new(1930502023) } }, // chaos barrier
@@ -381,7 +380,7 @@ internal static class JewelSpawnSystemPatch
         { new(-358319417), new() { new(281216122), new(-1310320536), new(-2083269917), new(1886458301), new(2113057383), new(681802645) } }, // void
         { new(1019568127), new() { new(1104681306), new(-681348970), new(2113057383), new(1439297485), new(-628722771), new(-2009288107) } }, // chaos volley
         { new(711231628), new() { new(2000559018), new(-593156502), new(-812464660), new(1702103303), new(255266111), new(-1430581265) } }, // veil of chaos
-        { new(61360182), new() { new(-1062097246), new(901495348), new(-1649014147), new(-689590392), new(-1910609332), new(321435434), new(-919146370), new(1401753271) } }, // rain of chaos
+        { new(2012523607), new() { new(-33694441), new(956784340), new(-305171830), new(2113057383), new(1439297485), new(1326058627) } }, // rain of chaos
         // frost
         { new(-1000260252), new() { new(1336836422), new(-1757583318), new(986977415), new(1616797198), new(-311910625), new(1222918506), new(291310353) } }, // cold snap
         { new(295045820), new() { new(-771579655), new(-30104212), new(-111114882), new(-311910625), new(950989548), new(-2009288107) } }, // crystal lance
@@ -389,7 +388,7 @@ internal static class JewelSpawnSystemPatch
         { new(78384915), new() { new(1644464649), new(440375591), new(-111114882), new(-2047023759), new(950989548), new(-2009288107) } }, // frost bat
         { new(91249849), new() { new(-1070941840), new(-1916056946), new(1934366532), new(1439297485), new(681802645) } }, // ice nova
         { new(1709284795), new() { new(-292495274), new(1126070097), new(-1378154439), new(620700670), new(255266111), new(-1430581265) } }, // veil of frost
-        { new(1119012588), new() { new(-1285185439), new(-1655280244), new(1298763769), new(701417059), new(855985036), new(-269336230), new(-16130732), new(320961855) } }, // frost cone
+        { new(1119012588), new() { new(287381106), new(1362931942), new(-1578113529), new(-1829520054), new(1471056038), new(-539833003), new(1476790178) } }, // frost cone
         // illusion
         { new(110097606), new() { new(-845453001), new(1301174222), new(-415768376), new(1891772829), new(1552774208), new(291310353), new(-1967899075), new(-1274845133) } }, // mist trance
         { new(268059675), new() { new(-529803606), new(1212582123), new(-1928057811), new(-1673859267), new(-1087850059) } }, // mosquito
@@ -397,15 +396,15 @@ internal static class JewelSpawnSystemPatch
         { new(247896794), new() { new(1531499726), new(1610681142), new(-2009288107), new(1499233761), new(-389780147), new(-1224808007), new(424876885), new(-191364711) } }, // spectral wolf
         { new(-242769430), new() { new(-1565427919), new(1531499726), new(1610681142), new(-1772665607), new(-1653068805), new(-233951066), new(-1538705520) } }, // wraith spear
         { new(-935015750), new() { new(2138408718), new(557219983), new(1016557168), new(-1743623080), new(255266111), new(-1430581265), new(-450361030) } }, // veil of illusion
-        { new(1210635702), new() { new(1142315183), new(-1317668918), new(2089605812), new(390853496), new(-988990419), new(1691986235), new(-1200048631), new(-730437928) } }, // curse
+        { new(-1432758970), new() { new(240097113), new(401599216), new(-1926752281), new(570496464), new(1881441820), new(1326058627), new(-1967899075) } }, // curse
         // storm
         { new(1249925269), new() { new(-531481445), new(353305817), new(-485022350), new(-316223882), new(-1772665607), new(292333199) } }, // ball lightning
-        { new(-356990326), new() { new(-1643437789), new(2062783787), new(-2009288107), new(1215957974), new(946721895) } }, // cyclone
+        { new(-356990326), new() { new(-1643437789), new(-1842616277), new(-703040254), new(2062783787), new(-2009288107), new(1215957974), new(946721895), new(-111114882), new(-1730057341) } }, // cyclone
         { new(1952703098), new() { new(171817139), new(-2071143392), new(98803150), new(1158616225), new(1113225149), new(291310353), new(-1202845465) } }, // discharge
         { new(1071205195), new() { new(1780108774), new(-928750139), new(-635781998), new(-743834336), new(-2109940363) } }, // lightning wall
         { new(-987810170), new() { new(-1565427919), new(-2009288107), new(946721895), new(958439837), new(578859494) } }, // polarity shift
         { new(-84816111), new() { new(1215957974), new(255266111), new(-1430581265), new(-387102419), new(-115293432), new(1221500964) } }, // veil of storm
-        { new(-1025530809), new() { new(-396872793), new(259467316), new(917990176), new(-2055591403), new(-437325088), new(-820686873), new(90716564), new(-703021474) } }, // lightning tendrils
+        { new(-1184139778), new() { new(723250184), new(1576597957), new(-1680074607), new(-402953124), new(1488885094), new(-2009288107), new(-111114882) } }, // lightning tendrils
         // unholy
         { new(481411985), new() { new(585605138), new(-968605931), new(1291379982), new(-612004637), new(47727933), new(419000172), new(1439297485), new(681802645) } }, // bone explosion
         { new(-1204819086), new() { new(1562979558), new(538792139), new(1944307151), new(-203019589), new(-1967899075), new(-2009288107) } }, // corrupted skull
@@ -413,7 +412,7 @@ internal static class JewelSpawnSystemPatch
         { new(2138402840), new() { new(-696735285), new(-1096014124), new(1670819844), new(-249390913), new(15549217), new(219517192), new(-770033390), new(1871790882) } }, // soulburn
         { new(-1136860480), new() { new(1930502023), new(-1729725919), new(1998410228), new(761541981), new(-649562549), new(909721987), new(-2133606415), new(-1840862497) } }, // ward of the damned
         { new(-498302954), new() { new(-319638993), new(-394612778), new(-1776361271), new(952126692), new(255266111), new(-1430581265) } }, // veil of bones
-        { new(-782050236), new() { new(1325686057), new(-931371902), new(1727243260), new(-471020436), new(-230762527), new(974552890), new(2041980895), new(-1968712128) } }, // chains of death
+        { new(-1845982676), new() { new(-1214230143), new(599900100), new(1558320813), new(-1709420734), new(-1529767682), new(1756722785), new(1235988685), new(1618961316), new(-1853462369), new(1549057800) } }, // chains of death
     };
 
     /*